@@ -9,23 +9,23 @@ dependencies:
     # The example app is bundled with the plugin so we use a path dependency on
     # the parent directory to use the current plugin's version.
     path: ../
-  path_provider: ^0.5.0
+  path_provider: ^2.0.0
   flutter:
     sdk: flutter
-  video_player: ^0.10.0
-  integration_test:
-    path: ../../../integration_test
+  video_player: ^2.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_driver:
     sdk: flutter
-  pedantic: ^1.8.0
+  integration_test:
+    path: ../../../integration_test
+  pedantic: ^1.10.0
 
 flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.9.1+hotfix.4"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
+  flutter: ">=1.22.0"
