@@ -33,4 +33,4 @@ dev_dependencies:
 
 environment:
   sdk: ">=2.3.0 <3.0.0"
-  flutter: ">=1.10.0"
+  flutter: ">=1.20.0"
