@@ -2,38 +2,33 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.9
-
 import 'dart:io';
 import 'dart:ui';
 
-import 'package:flutter_test/flutter_test.dart' show TestWidgetsFlutterBinding;
-import 'package:mockito/mockito.dart';
-import 'package:share/share.dart';
-import 'package:test/test.dart';
-
 import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:share/share.dart';
 
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
+  TestWidgetsFlutterBinding.ensureInitialized(); // Required for MethodChannels
 
-  MockMethodChannel mockChannel;
+  late FakeMethodChannel fakeChannel;
 
   setUp(() {
-    mockChannel = MockMethodChannel();
-    // Re-pipe to mockito for easier verifies.
+    fakeChannel = FakeMethodChannel();
+    // Re-pipe to our fake to verify invocations.
     Share.channel.setMockMethodCallHandler((MethodCall call) async {
       // The explicit type can be void as the only method call has a return type of void.
-      await mockChannel.invokeMethod<void>(call.method, call.arguments);
+      await fakeChannel.invokeMethod<void>(call.method, call.arguments);
     });
   });
 
   test('sharing empty fails', () {
     expect(
       () => Share.share(''),
-      throwsA(const TypeMatcher<AssertionError>()),
+      throwsA(isA<AssertionError>()),
     );
-    verifyZeroInteractions(mockChannel);
+    expect(fakeChannel.invocation, isNull);
   });
 
   test('sharing origin sets the right params', () async {
@@ -42,22 +37,28 @@ void main() {
       subject: 'some subject to share',
       sharePositionOrigin: const Rect.fromLTWH(1.0, 2.0, 3.0, 4.0),
     );
-    verify(mockChannel.invokeMethod<void>('share', <String, dynamic>{
-      'text': 'some text to share',
-      'subject': 'some subject to share',
-      'originX': 1.0,
-      'originY': 2.0,
-      'originWidth': 3.0,
-      'originHeight': 4.0,
-    }));
+
+    expect(
+      fakeChannel.invocation,
+      equals({
+        'share': {
+          'text': 'some text to share',
+          'subject': 'some subject to share',
+          'originX': 1.0,
+          'originY': 2.0,
+          'originWidth': 3.0,
+          'originHeight': 4.0,
+        }
+      }),
+    );
   });
 
   test('sharing empty file fails', () {
     expect(
       () => Share.shareFiles(['']),
-      throwsA(const TypeMatcher<AssertionError>()),
+      throwsA(isA<AssertionError>()),
     );
-    verifyZeroInteractions(mockChannel);
+    expect(fakeChannel.invocation, isNull);
   });
 
   test('sharing file sets correct mimeType', () async {
@@ -65,11 +66,18 @@ void main() {
     final File file = File(path);
     try {
       file.createSync();
+
       await Share.shareFiles([path]);
-      verify(mockChannel.invokeMethod('shareFiles', <String, dynamic>{
-        'paths': [path],
-        'mimeTypes': ['image/png'],
-      }));
+
+      expect(
+        fakeChannel.invocation,
+        equals({
+          'shareFiles': {
+            'paths': [path],
+            'mimeTypes': ['image/png'],
+          }
+        }),
+      );
     } finally {
       file.deleteSync();
     }
@@ -80,15 +88,30 @@ void main() {
     final File file = File(path);
     try {
       file.createSync();
+
       await Share.shareFiles([path], mimeTypes: ['*/*']);
-      verify(mockChannel.invokeMethod('shareFiles', <String, dynamic>{
-        'paths': [file.path],
-        'mimeTypes': ['*/*'],
-      }));
+
+      expect(
+        fakeChannel.invocation,
+        equals({
+          'shareFiles': {
+            'paths': [file.path],
+            'mimeTypes': ['*/*'],
+          }
+        }),
+      );
     } finally {
       file.deleteSync();
     }
   });
 }
 
-class MockMethodChannel extends Mock implements MethodChannel {}
+class FakeMethodChannel extends Fake implements MethodChannel {
+  Map<String, dynamic>? invocation;
+
+  @override
+  Future<T?> invokeMethod<T>(String method, [dynamic arguments]) async {
+    this.invocation = {method: arguments};
+    return null;
+  }
+}
