@@ -2,26 +2,26 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// TODO(egarciad): Remove once Mockito has been migrated to null safety.
-// @dart = 2.9
-
 import 'dart:ui';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:flutter/services.dart';
-import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 import 'package:url_launcher/link.dart';
 import 'package:url_launcher_platform_interface/url_launcher_platform_interface.dart';
+
+import 'mock_url_launcher_platform.dart';
 
 final MethodCodec _codec = const JSONMethodCodec();
 
 void main() {
-  final MockUrlLauncher mock = MockUrlLauncher();
-  UrlLauncherPlatform.instance = mock;
+  late MockUrlLauncher mock;
 
-  PlatformMessageCallback realOnPlatformMessage;
+  PlatformMessageCallback? realOnPlatformMessage;
   setUp(() {
+    mock = MockUrlLauncher();
+    UrlLauncherPlatform.instance = mock;
     realOnPlatformMessage = window.onPlatformMessage;
   });
   tearDown(() {
@@ -31,11 +31,11 @@ void main() {
   group('$Link', () {
     testWidgets('handles null uri correctly', (WidgetTester tester) async {
       bool isBuilt = false;
-      FollowLink followLink;
+      FollowLink? followLink;
 
       final Link link = Link(
         uri: null,
-        builder: (BuildContext context, FollowLink followLink2) {
+        builder: (BuildContext context, FollowLink? followLink2) {
           isBuilt = true;
           followLink = followLink2;
           return Container();
@@ -50,66 +50,62 @@ void main() {
 
     testWidgets('calls url_launcher for external URLs with blank target',
         (WidgetTester tester) async {
-      FollowLink followLink;
+      FollowLink? followLink;
 
       await tester.pumpWidget(Link(
         uri: Uri.parse('http://example.com/foobar'),
         target: LinkTarget.blank,
-        builder: (BuildContext context, FollowLink followLink2) {
+        builder: (BuildContext context, FollowLink? followLink2) {
           followLink = followLink2;
           return Container();
         },
       ));
 
-      when(mock.canLaunch('http://example.com/foobar'))
-          .thenAnswer((realInvocation) => Future<bool>.value(true));
-      clearInteractions(mock);
-      await followLink();
-
-      verifyInOrder([
-        mock.canLaunch('http://example.com/foobar'),
-        mock.launch(
-          'http://example.com/foobar',
+      mock
+        ..setLaunchExpectations(
+          url: 'http://example.com/foobar',
           useSafariVC: false,
           useWebView: false,
           universalLinksOnly: false,
           enableJavaScript: false,
           enableDomStorage: false,
           headers: <String, String>{},
+          webOnlyWindowName: null,
         )
-      ]);
+        ..setResponse(true);
+      await followLink!();
+      expect(mock.canLaunchCalled, isTrue);
+      expect(mock.launchCalled, isTrue);
     });
 
     testWidgets('calls url_launcher for external URLs with self target',
         (WidgetTester tester) async {
-      FollowLink followLink;
+      FollowLink? followLink;
 
       await tester.pumpWidget(Link(
         uri: Uri.parse('http://example.com/foobar'),
         target: LinkTarget.self,
-        builder: (BuildContext context, FollowLink followLink2) {
+        builder: (BuildContext context, FollowLink? followLink2) {
           followLink = followLink2;
           return Container();
         },
       ));
 
-      when(mock.canLaunch('http://example.com/foobar'))
-          .thenAnswer((realInvocation) => Future<bool>.value(true));
-      clearInteractions(mock);
-      await followLink();
-
-      verifyInOrder([
-        mock.canLaunch('http://example.com/foobar'),
-        mock.launch(
-          'http://example.com/foobar',
+      mock
+        ..setLaunchExpectations(
+          url: 'http://example.com/foobar',
           useSafariVC: true,
           useWebView: true,
           universalLinksOnly: false,
           enableJavaScript: false,
           enableDomStorage: false,
           headers: <String, String>{},
+          webOnlyWindowName: null,
         )
-      ]);
+        ..setResponse(true);
+      await followLink!();
+      expect(mock.canLaunchCalled, isTrue);
+      expect(mock.launchCalled, isTrue);
     });
 
     testWidgets('sends navigation platform messages for internal route names',
@@ -125,21 +121,21 @@ void main() {
       final List<MethodCall> frameworkCalls = <MethodCall>[];
       window.onPlatformMessage = (
         String name,
-        ByteData data,
-        PlatformMessageResponseCallback callback,
+        ByteData? data,
+        PlatformMessageResponseCallback? callback,
       ) {
         frameworkCalls.add(_codec.decodeMethodCall(data));
-        realOnPlatformMessage(name, data, callback);
+        realOnPlatformMessage!(name, data, callback);
       };
 
       final Uri uri = Uri.parse('/foo/bar');
-      FollowLink followLink;
+      FollowLink? followLink;
 
       await tester.pumpWidget(MaterialApp(
         routes: <String, WidgetBuilder>{
           '/': (BuildContext context) => Link(
                 uri: uri,
-                builder: (BuildContext context, FollowLink followLink2) {
+                builder: (BuildContext context, FollowLink? followLink2) {
                   followLink = followLink2;
                   return Container();
                 },
@@ -150,11 +146,11 @@ void main() {
 
       engineCalls.clear();
       frameworkCalls.clear();
-      clearInteractions(mock);
-      await followLink();
+      await followLink!();
 
       // Shouldn't use url_launcher when uri is an internal route name.
-      verifyZeroInteractions(mock);
+      expect(mock.canLaunchCalled, isFalse);
+      expect(mock.launchCalled, isFalse);
 
       // A message should've been sent to the engine (by the Navigator, not by
       // the Link widget).
@@ -191,19 +187,19 @@ void main() {
       final List<MethodCall> frameworkCalls = <MethodCall>[];
       window.onPlatformMessage = (
         String name,
-        ByteData data,
-        PlatformMessageResponseCallback callback,
+        ByteData? data,
+        PlatformMessageResponseCallback? callback,
       ) {
         frameworkCalls.add(_codec.decodeMethodCall(data));
-        realOnPlatformMessage(name, data, callback);
+        realOnPlatformMessage!(name, data, callback);
       };
 
       final Uri uri = Uri.parse('/foo/bar');
-      FollowLink followLink;
+      FollowLink? followLink;
 
       final Link link = Link(
         uri: uri,
-        builder: (BuildContext context, FollowLink followLink2) {
+        builder: (BuildContext context, FollowLink? followLink2) {
           followLink = followLink2;
           return Container();
         },
@@ -217,11 +213,11 @@ void main() {
 
       engineCalls.clear();
       frameworkCalls.clear();
-      clearInteractions(mock);
-      await followLink();
+      await followLink!();
 
       // Shouldn't use url_launcher when uri is an internal route name.
-      verifyZeroInteractions(mock);
+      expect(mock.canLaunchCalled, isFalse);
+      expect(mock.launchCalled, isFalse);
 
       // Sends route information update to the engine.
       expect(engineCalls, hasLength(1));
@@ -249,10 +245,6 @@ void main() {
   });
 }
 
-class MockUrlLauncher extends Mock
-    with MockPlatformInterfaceMixin
-    implements UrlLauncherPlatform {}
-
 class MockRouteInformationParser extends Mock
     implements RouteInformationParser<bool> {
   @override
@@ -261,8 +253,8 @@ class MockRouteInformationParser extends Mock
   }
 }
 
-class MockRouterDelegate extends Mock implements RouterDelegate {
-  MockRouterDelegate({@required this.builder});
+class MockRouterDelegate extends Mock implements RouterDelegate<Object> {
+  MockRouterDelegate({required this.builder});
 
   final WidgetBuilder builder;
 
@@ -270,4 +262,10 @@ class MockRouterDelegate extends Mock implements RouterDelegate {
   Widget build(BuildContext context) {
     return builder(context);
   }
+
+  @override
+  Future<void> setInitialRoutePath(Object configuration) async {}
+
+  @override
+  Future<void> setNewRoutePath(Object configuration) async {}
 }
