@@ -1,9 +1,7 @@
-import 'package:integration_test/integration_test.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:connectivity_platform_interface/connectivity_platform_interface.dart';
 import 'package:connectivity_for_web/src/network_information_api_connectivity_plugin.dart';
-
-import 'package:mockito/mockito.dart';
+import 'package:connectivity_platform_interface/connectivity_platform_interface.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:integration_test/integration_test.dart';
 
 import 'src/connectivity_mocks.dart';
 
@@ -12,66 +10,70 @@ void main() {
 
   group('checkConnectivity', () {
     void testCheckConnectivity({
-      String type,
-      String effectiveType,
-      num downlink = 10,
-      num rtt = 50,
-      ConnectivityResult expected,
+      String? type,
+      String? effectiveType,
+      num? downlink = 10,
+      int? rtt = 50,
+      required ConnectivityResult expected,
     }) {
-      final connection = MockNetworkInformation();
-      when(connection.type).thenReturn(type);
-      when(connection.effectiveType).thenReturn(effectiveType);
-      when(connection.downlink).thenReturn(downlink);
-      when(connection.rtt).thenReturn(downlink);
+      final connection = FakeNetworkInformation(
+        type: type,
+        effectiveType: effectiveType,
+        downlink: downlink,
+        rtt: rtt,
+      );
 
       NetworkInformationApiConnectivityPlugin plugin =
           NetworkInformationApiConnectivityPlugin.withConnection(connection);
       expect(plugin.checkConnectivity(), completion(equals(expected)));
     }
 
-    test('0 downlink and rtt -> none', () {
+    testWidgets('0 downlink and rtt -> none', (WidgetTester tester) async {
       testCheckConnectivity(
           effectiveType: '4g',
           downlink: 0,
           rtt: 0,
           expected: ConnectivityResult.none);
     });
-    test('slow-2g -> mobile', () {
+    testWidgets('slow-2g -> mobile', (WidgetTester tester) async {
       testCheckConnectivity(
           effectiveType: 'slow-2g', expected: ConnectivityResult.mobile);
     });
-    test('2g -> mobile', () {
+    testWidgets('2g -> mobile', (WidgetTester tester) async {
       testCheckConnectivity(
           effectiveType: '2g', expected: ConnectivityResult.mobile);
     });
-    test('3g -> mobile', () {
+    testWidgets('3g -> mobile', (WidgetTester tester) async {
       testCheckConnectivity(
           effectiveType: '3g', expected: ConnectivityResult.mobile);
     });
-    test('4g -> wifi', () {
+    testWidgets('4g -> wifi', (WidgetTester tester) async {
       testCheckConnectivity(
           effectiveType: '4g', expected: ConnectivityResult.wifi);
     });
   });
 
   group('get onConnectivityChanged', () {
-    test('puts change events in a Stream', () async {
-      final connection = MockNetworkInformation();
+    testWidgets('puts change events in a Stream', (WidgetTester tester) async {
+      final connection = FakeNetworkInformation();
       NetworkInformationApiConnectivityPlugin plugin =
           NetworkInformationApiConnectivityPlugin.withConnection(connection);
 
-      Stream<ConnectivityResult> results = plugin.onConnectivityChanged;
+      // The onConnectivityChanged stream is infinite, so we only .take(2) so the test completes.
+      // We need to do .toList() now, because otherwise the Stream won't be actually listened to,
+      // and we'll miss the calls to mockChangeValue below.
+      final results = plugin.onConnectivityChanged.take(2).toList();
 
       // Fake a disconnect-reconnect
       await connection.mockChangeValue(downlink: 0, rtt: 0);
       await connection.mockChangeValue(
           downlink: 10, rtt: 50, effectiveType: '4g');
 
-      // The stream of results is infinite, so we need to .take(2) for this test to complete.
+      // Expect to see the disconnect-reconnect in the resulting stream.
       expect(
-          results.take(2).toList(),
-          completion(
-              equals([ConnectivityResult.none, ConnectivityResult.wifi])));
+        results,
+        completion([ConnectivityResult.none, ConnectivityResult.wifi]),
+      );
     });
   });
 }
