@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'dart:html' as html show window, NetworkInformation;
-import 'dart:js';
-import 'dart:js_util';
+import 'dart:js' show allowInterop;
+import 'dart:js_util' show setProperty;
 
 import 'package:connectivity_platform_interface/connectivity_platform_interface.dart';
 import 'package:connectivity_for_web/connectivity_for_web.dart';
@@ -18,7 +18,7 @@ class NetworkInformationApiConnectivityPlugin extends ConnectivityPlugin {
 
   /// The constructor of the plugin.
   NetworkInformationApiConnectivityPlugin()
-      : this.withConnection(html.window.navigator.connection);
+      : this.withConnection(html.window.navigator.connection!);
 
   /// Creates the plugin, with an override of the NetworkInformation object.
   @visibleForTesting
@@ -32,8 +32,8 @@ class NetworkInformationApiConnectivityPlugin extends ConnectivityPlugin {
     return networkInformationToConnectivityResult(_networkInformation);
   }
 
-  StreamController<ConnectivityResult> _connectivityResultStreamController;
-  Stream<ConnectivityResult> _connectivityResultStream;
+  StreamController<ConnectivityResult>? _connectivityResultStreamController;
+  late Stream<ConnectivityResult> _connectivityResultStream;
 
   /// Returns a Stream of ConnectivityResults changes.
   @override
@@ -41,8 +41,10 @@ class NetworkInformationApiConnectivityPlugin extends ConnectivityPlugin {
     if (_connectivityResultStreamController == null) {
       _connectivityResultStreamController =
           StreamController<ConnectivityResult>();
+
+      // Directly write the 'onchange' function on the networkInformation object.
       setProperty(_networkInformation, 'onchange', allowInterop((_) {
-        _connectivityResultStreamController
+        _connectivityResultStreamController!
             .add(networkInformationToConnectivityResult(_networkInformation));
       }));
       // TODO: Implement the above with _networkInformation.onChange:
@@ -54,7 +56,7 @@ class NetworkInformationApiConnectivityPlugin extends ConnectivityPlugin {
       // onChange Stream upon hot restart.
       // https://github.com/dart-lang/sdk/issues/42679
       _connectivityResultStream =
-          _connectivityResultStreamController.stream.asBroadcastStream();
+          _connectivityResultStreamController!.stream.asBroadcastStream();
     }
     return _connectivityResultStream;
   }
