@@ -84,8 +84,8 @@ class AppStoreConnection implements InAppPurchaseConnection {
   }
 
   @override
-  Future<BillingResultWrapper> completePurchase(PurchaseDetails purchase,
-      {String? developerPayload}) async {
+  Future<BillingResultWrapper> completePurchase(
+      PurchaseDetails purchase) async {
     if (purchase.skPaymentTransaction == null) {
       throw ArgumentError(
           'completePurchase unsuccessful. The `purchase.skPaymentTransaction` is not valid');
@@ -96,8 +96,7 @@ class AppStoreConnection implements InAppPurchaseConnection {
   }
 
   @override
-  Future<BillingResultWrapper> consumePurchase(PurchaseDetails purchase,
-      {String? developerPayload}) {
+  Future<BillingResultWrapper> consumePurchase(PurchaseDetails purchase) {
     throw UnsupportedError('consume purchase is not available on Android');
   }
 
