@@ -66,6 +66,8 @@ class GooglePlayConnection
             accountId: purchaseParam.applicationUserName,
             oldSku: purchaseParam
                 .changeSubscriptionParam?.oldPurchaseDetails.productID,
+            purchaseToken: purchaseParam.changeSubscriptionParam
+                ?.oldPurchaseDetails.verificationData.serverVerificationData,
             prorationMode:
                 purchaseParam.changeSubscriptionParam?.prorationMode);
     return billingResultWrapper.responseCode == BillingResponse.ok;
@@ -81,8 +83,8 @@ class GooglePlayConnection
   }
 
   @override
-  Future<BillingResultWrapper> completePurchase(PurchaseDetails purchase,
-      {String? developerPayload}) async {
+  Future<BillingResultWrapper> completePurchase(
+      PurchaseDetails purchase) async {
     if (purchase.billingClientPurchase!.isAcknowledged) {
       return BillingResultWrapper(responseCode: BillingResponse.ok);
     }
@@ -90,21 +92,18 @@ class GooglePlayConnection
       throw ArgumentError(
           'completePurchase unsuccessful. The `purchase.verificationData` is not valid');
     }
-    return await billingClient.acknowledgePurchase(
-        purchase.verificationData.serverVerificationData,
-        developerPayload: developerPayload);
+    return await billingClient
+        .acknowledgePurchase(purchase.verificationData.serverVerificationData);
   }
 
   @override
-  Future<BillingResultWrapper> consumePurchase(PurchaseDetails purchase,
-      {String? developerPayload}) {
+  Future<BillingResultWrapper> consumePurchase(PurchaseDetails purchase) {
     if (purchase.verificationData == null) {
       throw ArgumentError(
           'consumePurchase unsuccessful. The `purchase.verificationData` is not valid');
     }
-    return billingClient.consumeAsync(
-        purchase.verificationData.serverVerificationData,
-        developerPayload: developerPayload);
+    return billingClient
+        .consumeAsync(purchase.verificationData.serverVerificationData);
   }
 
   @override
