@@ -196,15 +196,53 @@ void main() {
       );
       final SkuDetailsWrapper skuDetails = dummySkuDetails;
       final String accountId = "hashedAccountId";
+      final String profileId = "hashedProfileId";
 
       expect(
           await billingClient.launchBillingFlow(
-              sku: skuDetails.sku, accountId: accountId),
+              sku: skuDetails.sku,
+              accountId: accountId,
+              obfuscatedProfileId: profileId),
           equals(expectedBillingResult));
       Map<dynamic, dynamic> arguments =
           stubPlatform.previousCallMatching(launchMethodName).arguments;
       expect(arguments['sku'], equals(skuDetails.sku));
       expect(arguments['accountId'], equals(accountId));
+      expect(arguments['obfuscatedProfileId'], equals(profileId));
+    });
+
+    test(
+        'Change subscription throws assertion error `oldSku` and `purchaseToken` has different nullability',
+        () async {
+      const String debugMessage = 'dummy message';
+      final BillingResponse responseCode = BillingResponse.ok;
+      final BillingResultWrapper expectedBillingResult = BillingResultWrapper(
+          responseCode: responseCode, debugMessage: debugMessage);
+      stubPlatform.addResponse(
+        name: launchMethodName,
+        value: buildBillingResultMap(expectedBillingResult),
+      );
+      final SkuDetailsWrapper skuDetails = dummySkuDetails;
+      final String accountId = 'hashedAccountId';
+      final String profileId = 'hashedProfileId';
+
+      expect(
+          billingClient.launchBillingFlow(
+              sku: skuDetails.sku,
+              accountId: accountId,
+              obfuscatedProfileId: profileId,
+              oldSku: dummyOldPurchase.sku,
+              purchaseToken: null),
+          throwsAssertionError);
+
+      expect(
+          billingClient.launchBillingFlow(
+              sku: skuDetails.sku,
+              accountId: accountId,
+              obfuscatedProfileId: profileId,
+              oldSku: null,
+              purchaseToken: dummyOldPurchase.purchaseToken),
+          throwsAssertionError);
     });
 
     test(
@@ -219,19 +257,25 @@ void main() {
         value: buildBillingResultMap(expectedBillingResult),
       );
       final SkuDetailsWrapper skuDetails = dummySkuDetails;
-      final String accountId = "hashedAccountId";
+      final String accountId = 'hashedAccountId';
+      final String profileId = 'hashedProfileId';
 
       expect(
           await billingClient.launchBillingFlow(
               sku: skuDetails.sku,
               accountId: accountId,
-              oldSku: dummyOldPurchase.sku),
+              obfuscatedProfileId: profileId,
+              oldSku: dummyOldPurchase.sku,
+              purchaseToken: dummyOldPurchase.purchaseToken),
           equals(expectedBillingResult));
       Map<dynamic, dynamic> arguments =
           stubPlatform.previousCallMatching(launchMethodName).arguments;
       expect(arguments['sku'], equals(skuDetails.sku));
       expect(arguments['accountId'], equals(accountId));
       expect(arguments['oldSku'], equals(dummyOldPurchase.sku));
+      expect(
+          arguments['purchaseToken'], equals(dummyOldPurchase.purchaseToken));
+      expect(arguments['obfuscatedProfileId'], equals(profileId));
     });
 
     test(
@@ -246,21 +290,27 @@ void main() {
         value: buildBillingResultMap(expectedBillingResult),
       );
       final SkuDetailsWrapper skuDetails = dummySkuDetails;
-      final String accountId = "hashedAccountId";
+      final String accountId = 'hashedAccountId';
+      final String profileId = 'hashedProfileId';
       final prorationMode = ProrationMode.immediateAndChargeProratedPrice;
 
       expect(
           await billingClient.launchBillingFlow(
               sku: skuDetails.sku,
               accountId: accountId,
+              obfuscatedProfileId: profileId,
               oldSku: dummyOldPurchase.sku,
-              prorationMode: prorationMode),
+              prorationMode: prorationMode,
+              purchaseToken: dummyOldPurchase.purchaseToken),
           equals(expectedBillingResult));
       Map<dynamic, dynamic> arguments =
           stubPlatform.previousCallMatching(launchMethodName).arguments;
       expect(arguments['sku'], equals(skuDetails.sku));
       expect(arguments['accountId'], equals(accountId));
       expect(arguments['oldSku'], equals(dummyOldPurchase.sku));
+      expect(arguments['obfuscatedProfileId'], equals(profileId));
+      expect(
+          arguments['purchaseToken'], equals(dummyOldPurchase.purchaseToken));
       expect(arguments['prorationMode'],
           ProrationModeConverter().toJson(prorationMode));
     });
@@ -440,8 +490,8 @@ void main() {
           name: consumeMethodName,
           value: buildBillingResultMap(expectedBillingResult));
 
-      final BillingResultWrapper billingResult = await billingClient
-          .consumeAsync('dummy token', developerPayload: 'dummy payload');
+      final BillingResultWrapper billingResult =
+          await billingClient.consumeAsync('dummy token');
 
       expect(billingResult, equals(expectedBillingResult));
     });
@@ -451,8 +501,8 @@ void main() {
         name: consumeMethodName,
         value: null,
       );
-      final BillingResultWrapper billingResult = await billingClient
-          .consumeAsync('dummy token', developerPayload: 'dummy payload');
+      final BillingResultWrapper billingResult =
+          await billingClient.consumeAsync('dummy token');
 
       expect(
           billingResult,
@@ -475,8 +525,7 @@ void main() {
           value: buildBillingResultMap(expectedBillingResult));
 
       final BillingResultWrapper billingResult =
-          await billingClient.acknowledgePurchase('dummy token',
-              developerPayload: 'dummy payload');
+          await billingClient.acknowledgePurchase('dummy token');
 
       expect(billingResult, equals(expectedBillingResult));
     });
@@ -486,8 +535,7 @@ void main() {
         value: null,
       );
       final BillingResultWrapper billingResult =
-          await billingClient.acknowledgePurchase('dummy token',
-              developerPayload: 'dummy payload');
+          await billingClient.acknowledgePurchase('dummy token');
 
       expect(
           billingResult,
